refactor worker.jobs - completely broke maintenance mode
	modified:   worker/bridge_data.py
	new file:   worker/jobs.py
	deleted:    worker/jobs/__init__.py
	deleted:    worker/jobs/framework.py
	deleted:    worker/jobs/poppers.py
	deleted:    worker/jobs/scribe.py
	modified:   worker/workers/framework.py
	modified:   worker/workers/scribe.py
