+++ worker/bridge_data.py
@@ -17,7 +17,6 @@ class BridgeData:
     mutex = threading.Lock()
 
     def __init__(self):
-        #        super().__init__(args)
         random.seed()
         # I have to pass the args from the extended class, as the framework class doesn't
         # know what kind of polymorphism this worker is using
@@ -33,6 +32,7 @@ class BridgeData:
         self.api_key = os.environ.get("HORDE_API_KEY", "0000000000")
         self.max_threads = int(os.environ.get("HORDE_MAX_THREADS", 1))
         self.queue_size = int(os.environ.get("HORDE_QUEUE_SIZE", 0))
+
         # remove these for the refactor?
         self.stats_output_frequency = int(os.environ.get("STATS_OUTPUT_FREQUENCY", 30))
         self.disable_terminal_ui = os.environ.get("DISABLE_TERMINAL_UI", "false") == "true"
